# kernel: added attention bench for profiling before optimization