--- a/src/kernels/attention/CMakeLists.txt
+++ b/src/kernels/attention/CMakeLists.txt
@@ -1,9 +1,13 @@
+include(cc_binary)
 include(cc_library)
 include(cc_test)
 
 cc_library(
   NAME 
     attention.kernel
+  HDRS
+    attention_cpu.h
+    attention_kernel_sm80.cuh
   SRCS 
     # attention.cu
   DEPS
@@ -14,8 +18,6 @@ cc_library(
 cc_test(
   NAME
     attention_kernel_test
-  HDRS
-    attention_cpu.h
   SRCS
     cute_test.cpp
     attention_cpu_test.cpp
@@ -25,6 +27,18 @@ cc_test(
     glog::glog
     GTest::gtest_main
 )
+
+cc_binary(
+  NAME 
+    attention_sm80_bench
+  SRCS 
+    attention_sm80_bench.cu
+  DEPS
+    nvbench::nvbench
+    nvbench::main
+    :attention.kernel
+)
+
 
 
 add_subdirectory(flash_attn)